cmd/status: Refactor into own package
